Add stationary tank range ramp
Tanks now keep their existing 5-tile range while moving, then linearly ramp to 14 tiles after three stationary seconds without path-driven translation or hull rotation. Movement or hull pivoting resets the ramp to base range; turret aiming and external pushes do not.

Project the current tank weapon range to owner/allied snapshots via weaponRangeTiles, bump compact snapshots to v25, and update protocol/balance docs so selected range overlays can show the authoritative dynamic range.

Verification: cargo test --manifest-path server/Cargo.toml -p rts-sim tank_; cargo test --manifest-path server/Cargo.toml -p rts-sim anti_tank_gun_firing_reveal_lasts_for_firing_cycle_plus_half_second; cargo test --manifest-path server/Cargo.toml -p rts-protocol; node tests/protocol_parity.mjs; node tests/client_contracts/protocol_contracts.mjs; node tests/client_contracts/renderer_feedback_contracts.mjs; node scripts/check-faction-catalog-parity.mjs; node scripts/check-wiki.mjs

+++ server/crates/sim/src/game/services/combat/acquisition.rs
@@ -172,7 +172,7 @@ fn target_relevant_for_auto_acquisition(attacker: EntityKind, target: &Entity) -
 
 fn weapon_range_px(attacker: &Entity) -> f32 {
     let profile = effective_attack_profile(attacker);
-    profile.range_tiles as f32 * config::TILE_SIZE as f32 + attacker.radius() + super::RANGE_SLACK
+    profile.range_tiles * config::TILE_SIZE as f32 + attacker.radius() + super::RANGE_SLACK
 }
 
 #[allow(clippy::too_many_arguments)]
